move code to the main business class
